Use classifications from unclassified.csv
When raw data is being processed, any unclassified lines are exported to unclassified.csv. Type column is blank.

The user can then amend that file and enter classifications by hand for each transaction.  This change uses that (specified) data from unclassified.csv and updates classifications in classified.csv before raw data is processed.

It does not update categories.csv based on specified types in unclassified.csv. This is yet to be done.